+++ trunk/as_sublight/source/sublightUDP.h
@@ -22,7 +22,7 @@ class SublightUDP : public Sublight {
   private:
     SOCKET _sd;
     sockaddr_in dest_addr;
-    virtual void Send(uint64 data) const;
+    virtual void Send(uint32 data) const;
   public:
     SublightUDP(PClip child, const unsigned __int16 port, const char* const ip);
     virtual _stdcall ~SublightUDP();
